Combine the session of captures, manager and processors (#358)
* capture: implement the task wather

A task watcher waits for new tasks and returns the
task event, then the capture can start a processor
to handle the task.

It is a replacement of the ChangeFeedWatcher and
ProcessorWatcher, the new solution would be more
clear and simpler.
