+++ tests/_utils/start_tidb_cluster
@@ -242,20 +242,21 @@ max-open-files = 1000
 block-cache-size = "1GB"
 EOF
 
-echo "Starting Upstream TiFlash..."
-mkdir -p ${OUT_DIR}/tiflash/ && cp $CUR/tiflash-users.toml ${OUT_DIR}/tiflash/users.toml
-tiflash server --config-file "$OUT_DIR/tiflash-config.toml" &
-
-echo "Verifying Upsteam TiFlash is started..."
-# Make sure TiFlash is started.
-while ! curl -o /dev/null -sf http://127.0.0.1:17000/metrics 1>/dev/null 2>&1 ; do
-    i=$((i+1))
-    if [ "$i" -gt 10 ]; then
-        cat ${OUT_DIR}/tiflash/log/proxy.log
-        cat ${OUT_DIR}/tiflash/log/server.log
-        cat ${OUT_DIR}/tiflash/log/error.log
-        echo 'Failed to start TiFlash'
-        exit 1
-    fi
-    sleep 2
-done
+#echo "Starting Upstream TiFlash..."
+#mkdir -p ${OUT_DIR}/tiflash/ && cp $CUR/tiflash-users.toml ${OUT_DIR}/tiflash/users.toml
+#tiflash server --config-file "$OUT_DIR/tiflash-config.toml" &
+#
+#echo "Verifying Upsteam TiFlash is started..."
+## Make sure TiFlash is started.
+#while ! curl -o /dev/null -sf http://127.0.0.1:17000/metrics 1>/dev/null 2>&1 ; do
+#    i=$((i+1))
+#    if [ "$i" -gt 10 ]; then
+#        cat ${OUT_DIR}/tiflash/log/proxy.log
+#        cat ${OUT_DIR}/tiflash/log/server.log
+#        cat ${OUT_DIR}/tiflash/log/error.log
+#        echo 'Failed to start TiFlash'
+#        exit 1
+#    fi
+#    sleep 2
+#done
+#
